Struct linked_list::LinkedList

pub struct LinkedList<T> {
    // some fields omitted
}

An experimental rewrite of LinkedList to provide a more cursor-oriented API.

Methods

impl<T> LinkedList<T>

fn new() -> LinkedList<T>

Makes a new LinkedList.

fn push_back(&mut self, elem: T)

Appends an element to the back of the list.

fn push_front(&mut self, elem: T)

Appends an element to the front of the list.

fn pop_back(&mut self) -> Option<T>

Removes the element at back of the list. Returns None if the list is empty.

fn pop_front(&mut self) -> Option<T>

Removes the element at front of the list. Returns None if the list is empty.

fn front(&self) -> Option<&T>

Gets the element at the front of the list, or None if empty.

fn back(&self) -> Option<&T>

Gets the element at the back of the list, or None if empty.

fn front_mut(&mut self) -> Option<&mut T>

Gets the element at the front of the list mutably, or None if empty.

fn back_mut(&mut self) -> Option<&mut T>

Gets the element at the back of the list mutably, or None if empty.

fn insert(&mut self, index: usize, elem: T)

Inserts an element at the given index.

Panics

Panics if the index is greater than the length of the list.

fn remove(&mut self, index: usize) -> Option<T>

Removes the element at the given index. Returns None if the index is out of bounds.

fn split_at(&mut self, index: usize) -> LinkedList<T>

Splits the list into two lists at the given index. Returns the right side of the split. Returns an empty list if index is out of bounds.

fn append(&mut self, other: &mut LinkedList<T>)

Appends the given list to the end of this one. The old list will be empty afterwards.

fn splice(&mut self, index: usize, other: &mut LinkedList<T>)

Inserts the given list at the given index. The old list will be empty afterwards.

fn len(&self) -> usize

Gets the number of elements in the list.

fn is_empty(&self) -> bool

Whether the list is empty.

fn clear(&mut self)

Removes all elements from the list.

fn cursor(&mut self) -> Cursor<T>

Gets a cursor over the list.

fn iter<'a>(&'a self) -> Iter<'a, T>

Provides a forward iterator.

fn iter_mut<'a>(&'a mut self) -> IterMut<'a, T>

Provides a forward iterator with mutable references.

fn into_iter(self) -> IntoIter<T>

Consumes the list into an iterator yielding elements by value.