Crate linked_list

An alternative implementation of std::collections::LinkedList, featuring experimental Cursor-based APIs.

Structs

Cursor
A Cursor is like an iterator, except that it can freely seek back-and-forth, and can safely mutate the list during iteration. This is because the lifetime of its yielded references are tied to its own lifetime, instead of just the underlying list. This means cursors cannot yield multiple elements at once.

IntoIter
An iterator over mutable references to the items of a LinkedList.

Iter
An iterator over references to the items of a LinkedList.

IterMut
An iterator over mutable references to the items of a LinkedList.

LinkedList
An experimental rewrite of LinkedList to provide a more cursor-oriented API.