Trait linked_list::Wrapper

pub trait Wrapper<T: ?Sized> {
    // Required methods
    fn into_pointer(self) -> NonNull<T>;
    unsafe fn from_pointer(ptr: NonNull<T>) -> Self;
    fn as_ref(&self) -> &T;
}

Wraps an object to be inserted in a linked list.

Required Methods

fn into_pointer(self) -> NonNull<T>

Converts the wrapped object into a pointer that represents it.

unsafe fn from_pointer(ptr: NonNull<T>) -> Self

Converts the object back from the pointer representation.

Safety

The passed pointer must come from a previous call to Wrapper::into_pointer().

fn as_ref(&self) -> &T

Returns a reference to the wrapped object.

Object Safety

This trait is not object safe.